реализовано добавление/удаление игр в избранное и бронь

+++ backend/core/utils.py
@@ -8,13 +8,14 @@ from rest_framework import response, status
 from rest_framework.generics import get_object_or_404
 
 from consoles.models import Console
+from games.models import Game
 
 
-def add_and_del(add_serializer, model, request, console_id):
+def add_and_del_console(add_serializer, model, request, console_id):
     """Опция добавления и удаления консоли."""
     user = request.user
     data = {'user': user.id,
-            'recipe': console_id}
+            'console': console_id}
     serializer = add_serializer(data=data, context={'request': request})
     if request.method == 'POST':
         serializer.is_valid(raise_exception=True)
@@ -23,6 +24,23 @@ def add_and_del(add_serializer, model, request, console_id):
                                  status=status.HTTP_201_CREATED)
     get_object_or_404(
         model, user=user, console=get_object_or_404(Console, id=console_id)
+    ).delete()
+    return response.Response(status=status.HTTP_204_NO_CONTENT)
+
+
+def add_and_del_game(add_serializer, model, request, game_id):
+    """Опция добавления и удаления игры."""
+    user = request.user
+    data = {'user': user.id,
+            'game': game_id}
+    serializer = add_serializer(data=data, context={'request': request})
+    if request.method == 'POST':
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return response.Response(serializer.data,
+                                 status=status.HTTP_201_CREATED)
+    get_object_or_404(
+        model, user=user, game=get_object_or_404(Game, id=game_id)
     ).delete()
     return response.Response(status=status.HTTP_204_NO_CONTENT)
